added a select_from_table function
